# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API sistem rikub",
        "title": "Rikub Backend",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/pekerja": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint untuk mengambil semua pekerja",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerja"
                ],
                "summary": "GetAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PekerjaBase"
                                            }
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.InternalErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint untuk update data pekerja",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerja"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "description": "Data update pekerja",
                        "name": "update",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.PekerjaUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PekerjaBase"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.InternalErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint untuk menambah data pekerja",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerja"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "description": "Data pekerja",
                        "name": "pekerja",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PekerjaCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CreationSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " data": {
                                            "$ref": "#/definitions/dto.PekerjaBase"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.InternalErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/pekerja/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint untuk mengambil detail pekerja",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerja"
                ],
                "summary": "GetDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Public ID anak",
                        "name": "public_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " data": {
                                            "$ref": "#/definitions/dto.PekerjaBase"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.InternalErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/pekerja/pagination": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint untuk paginasi data pekerja",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pekerja"
                ],
                "summary": "Pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cursor untuk data sebelumnya",
                        "name": "before",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor untuk data berikutnya",
                        "name": "after",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PaginationDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.InternalErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " error": {
                                            "type": "string"
                                        },
                                        " message": {
                                            "type": "string"
                                        },
                                        " status_code": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DeviceBase": {
            "type": "object",
            "properties": {
                "mac_address": {
                    "type": "string"
                },
                "pekerja_id": {
                    "type": "integer"
                },
                "public_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/enum.DeviceStatus"
                },
                "telemetry": {
                    "description": "ListAlert []AlertBase `json:\"daftar_alert\"`",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TelemetryBase"
                    }
                }
            }
        },
...
        "dto.PaginationDTO": {
            "type": "object",
            "properties": {
                "data": {},
                "has_next": {
                    "type": "boolean"
                },
                "has_previous": {
                    "type": "boolean"
                },
                "next_cursor": {
                    "type": "string"
                },
                "previous_cursor": {
                    "type": "string"
                }
            }
        },
        "dto.PekerjaBase": {
            "type": "object",
            "properties": {
                "device": {
                    "$ref": "#/definitions/dto.DeviceBase"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "pengawas_id": {
                    "type": "integer"
                },
                "public_id": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                }
            }
        },
        "dto.PekerjaCreate": {
            "type": "object",
            "properties": {
                "device_public_id": {
                    "description": "PengawasPublicID uuid.UUID `json:\"pengawas_public_id\"`",
                    "type": "string"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                }
            }
        },
        "dto.PekerjaUpdate": {
            "type": "object",
            "properties": {
                "device_public_id": {
                    "type": "string"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                }
            }
        },
...
        "enum.DeviceStatus": {
            "type": "string",
            "enum": [
                "aktif",
                "tidak aktif"
            ],
            "x-enum-varnames": [
                "Aktif",
                "TidakAktif"
            ]
        },
...
        "response.BadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Gagal parsing request"
                },
                "status": {
                    "type": "string",
                    "example": "400 Bad Request"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "response.CreationSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Sukses menambahkan data"
                },
                "status": {
                    "type": "string",
                    "example": "201 Created"
                },
                "status_code": {
                    "type": "integer",
                    "example": 201
                }
            }
        },
        "response.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "message": {
                    "type": "string",
                    "example": "Kesalahan pada sisi server"
                },
                "status": {
                    "type": "string",
                    "example": "500 Internal Server Error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Request berhasil"
                },
                "status": {
                    "type": "string",
                    "example": "200 OK"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
    },
...
}